In mobility, fetch country geojson and mobility dates simultaneously before fetching mobility for latest date
